Initial commit for wrapping auth. Additionally, district-cases have been implemented
