+++ config/config.ini
@@ -1,0 +1,3 @@
+[CREDENTIALS]
+CLIENT_ID=
+CLIENT_SECRET=
